+++ lib/appium_lib_core/common/base/driver.rb
@@ -239,22 +239,6 @@ module Appium
         end
         alias is_keyboard_shown keyboard_shown?
 
-        # [DEPRECATION]
-        # Send keys for a current active element
-        # @param [String] key Input text
-        #
-        # @example
-        #
-        #   @driver.send_keys 'happy testing!'
-        #
-        def send_keys(*key)
-          ::Appium::Logger.warn(
-            '[DEPRECATION] Driver#send_keys is deprecated in W3C spec. Use driver.action.<command>.perform instead'
-          )
-          @bridge.send_keys_to_active_element(key)
-        end
-        alias type send_keys
-
         # Returns an instance of DriverSettings to call get/update.
         #
         # @example